Validation, First Draft

+++ Domain/src/main/java/team_f/domain/helper/IntegerHelper.java
@@ -9,7 +9,7 @@ public class IntegerHelper {
         return false;
     }
 
-    public static boolean isPositiveDefaultPoint(int points) {
+    public static boolean isPositiveDefaultPoint(double points) {
         if(points >= 0) {
             return true;
         }
